training/requirements.txt
fireworks-ai
pydantic
python-dotenv
requests
